// admin-ui/src/generated-api-types/models/OrganizerInfo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Object containing information about the event organizer.
 */
export type OrganizerInfo = {
    /**
     * Name of the organizer.
     */
    name: string;
    /**
     * Contact email address for the organizer. Optional.
     */
    contact_email?: string;
    /**
     * Website URL for the organizer. Optional.
     */
    website?: string;
};