// admin-ui/src/generated-api-types/models/ActionLink.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * An object representing a primary call-to-action link for the event. Optional.
 */
export type ActionLink = {
    url: string;
    text: string;
    /**
     * Type of action (e.g., "purchase", "rsvp", "register", "info"). Optional.
     */
    type?: string;
};